Keystone: fix dependencies
Keystone needs a few things to run.
Were these missing because they are pulled in by something else?

Also, stop python-keystoneclient from blocking keystone.
I assume that it's probably a good idea not to have the server and
the client on the same machine, but is there an actual problem
of having the two together? (e.g. file conflicts)

I'll try it anyway, because I need to test that this thing works and I
cba to boot up a new machine.